pyproject.toml
[tool.poetry]
name = "monitorias-backend"
version = "0.1.0"
description = ""
authors = ["Juan Monteiro <juan.monteiro@jmonteiro.net>"]

[tool.poetry.dependencies]
python = "^3.12"
Django = "^4.1"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.24.2"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
rollbar = "^0.16.3"
drf-access-policy = "^1.3.0"
django-filter = "^22.1"
django-crispy-forms = "^1.14.0"
drf-writable-nested = "^0.7.0"
marshmallow = "^3.19.0"
bleach = "^5.0.1"
isort = "5.12.0"
pillow = "^10.0.0"
django-cloudinary-storage = "^0.3.0"
setuptools = ">=3.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
pylint = "~2.14.4"
pylint-django = "^2.5.3"
isort = "^5.10.1"
tomli = "^2.0.1"

[tool.pylint]
load-plugins = ["pylint_django"]
django-settings-module= "monitorias.settings"
ignore-paths = [".*/migrations/.*$"]
disable = ["too-many-ancestors", "duplicate-code"]
good-names = [
    "e",
    "pk",
]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"